requirements.txt
# Django Framework
Django>=5.2.0,<6.0.0

# Environment Variables
django-environ>=0.11.2

# Production Server
gunicorn>=21.2.0
whitenoise>=6.6.0

# Database Drivers (optional - uncomment if needed)
# psycopg2-binary>=2.9.9  # PostgreSQL
# mysqlclient>=2.2.0      # MySQL

# Caching (optional - uncomment if needed)
# django-redis>=5.4.0
# redis>=5.0.1

# API & Serialization
djangorestframework>=3.14.0

# Image Processing (optional - uncomment if needed)
# Pillow>=10.1.0

# Testing & Quality
pytest>=7.4.3
pytest-django>=4.7.0
pytest-cov>=4.1.0
coverage>=7.3.4

# Code Quality Tools
black>=23.12.1
isort>=5.13.2
flake8>=7.0.0
pylint>=3.0.3
pylint-django>=2.5.5

# Pre-commit Hooks
pre-commit>=3.6.0

# Development Tools
django-extensions>=3.2.3
ipython>=8.19.0

# Security
django-cors-headers>=4.3.1
